UAST: split KotlinUastResolveProviderService into a separate file
